chore: rename appeal for clarity
