+++ lpa-submissions-e2e-tests/cypress/integration/appealDetailsSidebar/appealDetailsSidebarSteps.js
@@ -1,5 +1,5 @@
 import { Given, Then } from 'cypress-cucumber-preprocessor/steps';
-const {appeal} = require('../../fixtures/anAppeal.json');
+const appeal = require('../../fixtures/completedAppeal.json');
 
 Given('A subsection page is presented with id of {string}', (id) => {
   cy.goToPage('other-appeals', id);
