+++ lpa-submissions-e2e-tests/cypress/integration/common/checkYourAnswers.js
@@ -3,7 +3,7 @@ import { getSubTaskInfo } from '../../support/common/subTasks';
 
 const pageId = 'confirm-answers'
 
-const preCannedAppeal = require('../../fixtures/anAppeal.json');
+const appeal = require('../../fixtures/completedAppeal.json');
 
 let currentSubTask = {};
 
@@ -12,7 +12,7 @@ Before(() => {
 });
 
 Given('a change to answer {string} is requested from Change your answers page', (answer) => {
-  cy.insertAppealAndCreateReply(preCannedAppeal.appeal).as('appealReply');
+  cy.insertAppealAndCreateReply(appeal).as('appealReply');
   cy.get('@appealReply').then( (appealReply) => {
     cy.goToCheckYourAnswersPage(appealReply.appealId);
     currentSubTask = getSubTaskInfo(answer);
