+++ lpa-submissions-e2e-tests/cypress/integration/common/common.js
@@ -1,6 +1,6 @@
 import { Given, When } from 'cypress-cucumber-preprocessor/steps';
 
-const {appeal} = require('../../fixtures/anAppeal.json');
+const appeal = require('../../fixtures/completedAppeal.json');
 const reply = require('../../fixtures/completedAppealReply.json');
 
 Given('all the mandatory questions for the questionnaire have been completed', () => {
